requirements.txt
torch>=1.7.0
torchvision>=0.8.1
mxnet>=1.7.0.post1
scipy>=1.5.1
numpy>=1.19.4
scikit_image>=0.17.2
pandas>=1.0.5
opencv_python>=4.3.0.36
matplotlib>=3.3.0
Pillow>=8.1.0
tensorboardX>=2.1
tqdm
scikit-image
cython
-e git+http://github.com/lucasb-eyer/pydensecrf.git#egg=pydensecrf
